requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.0.0=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
blas=1.0=mkl
ca-certificates=2023.12.12=haa95532_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
charset-normalizer=3.3.2=pypi_0
colorama=0.4.6=pyhd8ed1ab_0
comm=0.1.4=pyhd8ed1ab_0
contourpy=1.2.0=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py39hd77b12b_0
decorator=5.1.1=pyhd8ed1ab_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_0
executing=2.0.1=pyhd8ed1ab_0
flatbuffers=23.5.26=pypi_0
fonttools=4.45.1=pypi_0
gast=0.4.0=pypi_0
google-auth=2.23.4=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.59.3=pypi_0
h5py=3.10.0=pypi_0
icc_rt=2022.1.0=h6049295_2
idna=3.6=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib-resources=6.1.1=pypi_0
importlib_metadata=6.8.0=hd8ed1ab_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.26.0=pyha63f2e9_0
ipython=8.18.1=pyh5737063_1
jedi=0.19.1=pyhd8ed1ab_0
joblib=1.2.0=py39haa95532_0
jupyter_client=8.6.0=pyhd8ed1ab_0
jupyter_core=4.12.0=py39hcbf5309_0
keras=2.10.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.5=pypi_0
libclang=16.0.6=pypi_0
libsodium=1.0.18=h8d14728_1
markdown=3.5.1=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.8=py39h2bbff1b_0
mkl_random=1.2.4=py39h59b6b97_0
nest-asyncio=1.5.8=pyhd8ed1ab_0
numpy=1.26.2=pypi_0
numpy-base=1.26.3=py39h65a83cf_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.1.78=pypi_0
openpyxl=3.1.2=pypi_0
openssl=3.0.12=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=23.2=pyhd8ed1ab_0
pandas=2.1.4=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.1.0=pypi_0
pip=23.3.1=py39haa95532_0
prompt-toolkit=3.0.41=pyha770c72_0
protobuf=3.19.6=pypi_0
psutil=5.9.0=py39h2bbff1b_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pydot=1.4.2=pypi_0
pygments=2.17.2=pyhd8ed1ab_0
pyparsing=3.1.1=pypi_0
python=3.9.16=h1aa4202_3
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2023.3.post1=pypi_0
pywin32=227=py39hb82d6ee_1
pyzmq=23.2.1=py39he46f08e_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.3.0=py39h4ed8f06_0
scipy=1.11.4=pypi_0
seaborn=0.13.0=pypi_0
setuptools=68.0.0=py39haa95532_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.6.2=pyhd8ed1ab_0
tbb=2021.8.0=h59b6b97_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.10.1=pypi_0
tensorflow-estimator=2.10.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tornado=6.2=py39hb82d6ee_0
traitlets=5.14.0=pyhd8ed1ab_0
typing_extensions=4.8.0=pyha770c72_0
tzdata=2023.3=pypi_0
urllib3=2.1.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.12=pyhd8ed1ab_0
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py39haa95532_0
wrapt=1.16.0=pypi_0
zeromq=4.3.4=h0e60522_1
zipp=3.17.0=pyhd8ed1ab_0
